New window to add annotations
