+++ vcell-client/src/main/java/cbit/vcell/client/desktop/biomodel/BioModelEditor.java
@@ -24,6 +24,7 @@ import java.util.Objects;
 import javax.swing.JComponent;
 import javax.swing.ListSelectionModel;
 
+import cbit.vcell.client.desktop.biomodel.annotations.AnnotationsPanel;
 import org.vcell.model.rbm.MolecularType;
 import org.vcell.model.rbm.SpeciesPattern;
 import org.vcell.pathway.BioPaxObject;
@@ -53,7 +54,6 @@ import cbit.vcell.client.task.ClientTaskDispatcher;
 import cbit.vcell.clientdb.DocumentManager;
 import cbit.vcell.data.DataSymbol;
 import cbit.vcell.desktop.BioModelNode;
-import cbit.vcell.desktop.BioModelNode.PublicationInfoNode;
 import cbit.vcell.geometry.CSGObject;
 import cbit.vcell.geometry.GeometryInfo;
 import cbit.vcell.geometry.gui.CSGObjectPropertiesPanel;
